+++ eli5/formatters/text.py
@@ -2,6 +2,7 @@
 from __future__ import absolute_import
 from itertools import chain
 import six
+from tabulate import tabulate
 from typing import List, Optional, Iterator
 
 from eli5.base import Explanation, FeatureImportances
@@ -9,7 +10,8 @@ from . import fields
 from .features import FormattedFeatureName
 from .utils import (
     format_signed, format_value, format_weight, has_any_values_for_weights,
-    replace_spaces, should_highlight_spaces, tabulate)
+    replace_spaces, should_highlight_spaces)
+from .utils import tabulate as eli5_tabulate
 from .trees import tree2text
 
 
@@ -106,7 +108,7 @@ def _method_lines(explanation):
 
 def _description_lines(explanation):
     # type: (Explanation) -> List[str]
-    return [explanation.description]
+    return [explanation.description or '']
 
 
 def _error_lines(explanation):
@@ -117,6 +119,7 @@ def _error_lines(explanation):
 def _feature_importances_lines(explanation, hl_spaces):
     # type: (Explanation, Optional[bool]) -> Iterator[str]
     max_width = 0
+    assert explanation.feature_importances is not None
     for line in _fi_lines(explanation.feature_importances, hl_spaces):
         max_width = max(max_width, len(line))
         yield line
@@ -146,13 +149,14 @@ def _fi_lines(feature_importances, hl_spaces):
 
 def _decision_tree_lines(explanation):
     # type: (Explanation) -> List[str]
+    assert explanation.decision_tree is not None
     return ["", tree2text(explanation.decision_tree)]
 
 
 def _transition_features_lines(explanation):
     # type: (Explanation) -> List[str]
-    from tabulate import tabulate  # type: ignore
     tf = explanation.transition_features
+    assert tf is not None
     return [
         "",
         "Transition features:",
@@ -169,7 +173,7 @@ def _targets_lines(explanation,  # type: Explanation
                    ):
     # type: (...) -> List[str]
     lines = []
-
+    assert explanation.targets is not None
     for target in explanation.targets:
         scores = _format_scores(target.proba, target.score)
         if scores:
@@ -199,7 +203,8 @@ def _targets_lines(explanation,  # type: Explanation
             col_align = 'rl'
 
         w = target.feature_weights
-        table = tabulate(
+        assert w is not None
+        table = eli5_tabulate(
             [table_line(fw) for fw in chain(w.pos, reversed(w.neg))],
             header=table_header,
             col_align=col_align,