+++ eli5/sklearn/transform.py
@@ -1,14 +1,13 @@
 """transform_feature_names implementations for scikit-learn transformers
 """
 
-import numpy as np  # type: ignore
-from sklearn.pipeline import Pipeline, FeatureUnion  # type: ignore
-from sklearn.feature_selection.base import SelectorMixin  # type: ignore
-from sklearn.linear_model import (  # type: ignore
-    RandomizedLogisticRegression,
-    RandomizedLasso,
-)
-from sklearn.preprocessing import (  # type: ignore
+import numpy as np
+from sklearn.pipeline import Pipeline, FeatureUnion
+try:
+    from sklearn.feature_selection import SelectorMixin
+except ImportError:
+    from sklearn.feature_selection.base import SelectorMixin
+from sklearn.preprocessing import (
     MinMaxScaler,
     StandardScaler,
     MaxAbsScaler,
@@ -22,13 +21,21 @@ from eli5.sklearn.utils import get_feature_names as _get_feature_names
 # Feature selection:
 
 @transform_feature_names.register(SelectorMixin)
-@transform_feature_names.register(RandomizedLogisticRegression)
-@transform_feature_names.register(RandomizedLasso)
 def _select_names(est, in_names=None):
     mask = est.get_support(indices=False)
     in_names = _get_feature_names(est, feature_names=in_names,
                                   num_features=len(mask))
     return [in_names[i] for i in np.flatnonzero(mask)]
+
+try:
+    from sklearn.linear_model import (
+        RandomizedLogisticRegression,
+        RandomizedLasso,
+    )
+    _select_names = transform_feature_names.register(RandomizedLasso)(_select_names)
+    _select_names = transform_feature_names.register(RandomizedLogisticRegression)(_select_names)
+except ImportError:     # Removed in scikit-learn 0.21
+    pass
 
 
 # Scaling