+++ tests/test_sklearn_transform.py
@@ -15,11 +15,18 @@ from sklearn.feature_selection import (
     RFECV,
     SelectFromModel,
 )
-from sklearn.linear_model import (
-    LogisticRegression,
-    RandomizedLogisticRegression,
-    RandomizedLasso,  # TODO: add tests and document
-)
+from sklearn.linear_model import LogisticRegression
+_additional_test_cases = []
+try:
+    from sklearn.linear_model import (  # type: ignore
+        RandomizedLogisticRegression,
+        RandomizedLasso,  # TODO: add tests and document
+    )
+    _additional_test_cases.append(
+        (RandomizedLogisticRegression(random_state=42),
+         ['<NAME1>', '<NAME2>', '<NAME3>']))
+except ImportError:     # Removed in scikit-learn 0.21
+    pass
 from sklearn.preprocessing import (
     MinMaxScaler,
     StandardScaler,
@@ -74,23 +81,21 @@ def selection_score_func(X, y):
     (VarianceThreshold(1.0), ['<NAME2>']),
     (GenericUnivariateSelect(), ['<NAME2>']),
     (GenericUnivariateSelect(mode='k_best', param=2), ['<NAME2>', '<NAME3>']),
-    (SelectFromModel(LogisticRegression('l1', C=0.01, random_state=42)),
+    (SelectFromModel(LogisticRegression('l1', C=0.01, solver='liblinear', random_state=42, multi_class='ovr')),
      ['<NAME0>', '<NAME2>']),
     (SelectFromModel(
         PermutationImportance(
-            LogisticRegression(random_state=42),
+            LogisticRegression(solver='liblinear', random_state=42),
             cv=5, random_state=42, refit=False,
         ),
         threshold=0.1,
      ),
      ['<NAME2>', '<NAME3>']),
-    (RFE(LogisticRegression(random_state=42), 2),
+    (RFE(LogisticRegression(solver='liblinear', random_state=42, multi_class='ovr'), 2),
      ['<NAME1>', '<NAME3>']),
-    (RFECV(LogisticRegression(random_state=42)),
+    (RFECV(LogisticRegression(solver='liblinear', random_state=42, multi_class='ovr'), cv=3),
      ['<NAME0>', '<NAME1>', '<NAME2>', '<NAME3>']),
-    (RandomizedLogisticRegression(random_state=42),
-     ['<NAME1>', '<NAME2>', '<NAME3>']),
-])
+] + _additional_test_cases)
 def test_transform_feature_names_iris(transformer, expected, iris_train):
     X, y, _, _ = iris_train
     transformer.fit(X, y)